py_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
openai==1.3.0
python-jose==3.3.0
requests==2.31.0
aiohttp==3.9.1 